Info on password email reset on localhost
Hi :)

Referring to this post: https://forum.getkirby.com/t/testing-email-forms-with-kirby

Perhaps (for the lesser advanced developers using Kirby) it would be nice with a infobox:
"Using a local development environment?" to inform that:

"Testing with private email from localhost can be difficult, as mails often don’t get through. Consider using a mail provider for testing." - or something alike.